[str] String API optimizations
Simplify the argument checks of some string functions. This is in
preparation of the following commits that will move these checks
directly into the encoding functions, and switch a lot of Parrot code
to the new string macros.

With the exception of STRING_length and STRING_byte_length, the string
macros don't check the string argument for NULL pointers. So the
following commits also fix some places that still didn't use
STRING_IS_NULL and STRINGNULL.

git-svn-id: https://svn.parrot.org/parrot/branches/string_macros@49220 d31e2699-5ff4-0310-a27c-f18f2fbe73fe